MAINT: Extract SpotsMap component
